Tab-completion added for in-game whispers.
In the game lobby you can tab-complete ANY nick, both from the lobby or from a match running. While playing you can only tab-complete ingame players and observers.
This patch works in the match, it helps answering whispers from people in lobby or in other games.
The nick will be added to a string (in the same way as observers already work) and get recalled with TAB key.