fix preferences dialog for loadgame hotkey
Some hotkeys (like the loadgame hotkey) are present in some, but not all hotkey scopes and teh hotkey preferences dialog was unable to handle that. One of the results was that it was not possible to assign the same hotkey to the loadgame hotkey and the loadmap hotkey, this commit attempts to fix that.