Fix bug #23908 - Crash in SDL_BlitSurface caused by integer underflow
The font rendering code under SDL2 seems to deal poorly with silly
values being passed to it as a cause of integer over/underflows. This
commit changes the types used in gui1 from unsigned to signed for
width and height. This causes a number of statements that previously
inadvertedly did unsigned arithmetics without over/underflow checking
to be doing signed arithmetics instead, which is probably what the
author intended.