Peek for redraw events after each call to SDL_PollEvent.
Since SDL_PollEvent internally pumps events there is is a chance that there will be a redraw event added to the SDL event queue, which could cause a crash if there is a game DRAW event before it in the event queue. This should fix it, but it's not the prettiest of fix and should be rewritten if it works to instead put draw events to the start of the internal event queue after it's been drained.