Disable Planned Recruits (Bug #24022)
Bug #24022 highlights one of many issues with planned recruits. gfgtdf suggests it might be worthwhile to simply disable recruits in planning mode.

This change will no longer allow fake recruits, but a normal recruit will still take place on attempting to do so during planning mode. It can be applied to 1.12 as well, if accepted, but I can understand if it is not as this is not an ideal solution.