Bump version
Yes, RC_VERSION_REVISION is 3 rather than 2. Because the Win32 version
info resource doesn't take text suffixes in the field where this value
is used, and a fourth number doesn't make sense for our usage, from now
on RC_VERSION_REVISION is one release ahead in order to clearly
distinguish previous release builds from in-dev builds, which are for
this purpose equivalent to the next/future release builds.