Work around crash/assertion caused by shortening filenames in gui1
Theis introduces the get_max_height and get_max_width methods in gui1
menu elements. The get_max_width method is then used to check against
the magic number -1 which is used to specify that there is no max
width for the menu.
If there is a max width then the old way of doing
things will be invoked.