use enums instead of strings for controllers in wesnothd::game
This is not only faster, it is also easier to understand, becasue it is
now obvious that every enement of side_controllers_ can only be 'human',
'ai' or 'null'.

This also makes enum::parse() in MAKE_ENUM a template. This way it also works
with other string classes, specially string_span from simple_wml.