SDL2: Set a minimum window size
This sets the minimum window size properly, allowing SDL to hint to
the window manager that the window should not be allowed to be changed
into a size smaller than the set size. This removes the need to
programatically try set the window size on a resize event if the new
size is smaller than the minimum size.