Fix bug #23934: Add event filtering on resize events
This adds event filtering and a timeout on actions after a resize
event. All DRAW, TOOLTIP and window events are filtered for 100ms
after a resize event has been received. The last received resize event
is injected into the event handling code when the next event is
received after the timeout has expired. This ensures that no more
resize events are dropped and that it's a better user experience.