Fixed crash on startup on Windows in paths with spaces
If OMP_WAIT_POLICY is not set, the game relaunches itself using
_execv() after setting the environment variable. It passes the entire
argv array as argv for the new launch, including argv[0] (program
path). If the program path contains spaces and is not surrounded in
quotes, Windows splits the path on quotes and creates a separate
parameter for each part (e.g.
"I:\Battle for Wesnoth\wesnoth\wesnoth.exe" becomes {"I:\Battle",
"for", "Wesnoth\wesnoth\wesnoth.exe"}). Boost::program_options
interprets the last two as positional parameters (the first one is
skipped, as the game expects the first parameter to store the program
path) and crashes the game because only one positional parameter is
allowed.

Fixed by quoting the path in the argv array.