Make the story screen resize aware
These changes make the storyscreen resize aware by adding an event
handler to the UI component and giving it its own event context. On an
window-event it'll restart relayout the entire scene and re-start the
drawing of the text.

Known issue: Gamescreen buttons appear when you move over them. This is beyond the scope of this commit and will be fixed in other changes.