Attempt to fix listbox scrollbar position after content changes.
Proviously, if a row was added/removed from a listbox, the scrolling
position of the listbox would be reset to 0. The expected behaviour is
that the visible part of the list does not change, this means that the
position is increased by the added space if the space was inserted
before the visible area, othewise it shouldn't be increased.

Unforunateley the code that adds/removes new rows does not know where
the extra space was inserted/removed because this is not calculated
until the next draw which calls layout_children(). So this doesn't work
very well, even after this commit.

Will try to fix this for treeviews later which might be easier because
treeviews don't use the tgenterator class and specially don't support
sorting of its elements.

We might want to revert the listbox.cpp/hpp part of this commit later
in case it doesn't work.