fix sides with no type= attribute
If you dont want that the engine uses the [side] wml table to create a
leader for that side, it is now enough to just omit the type attribute.
Previously it was necessary to add a no_leader=yes attribute.

The no_leader attribute is still used by the mp connect engine (which is
also used in sp) to prevent the mp connect engine from adding a random
leader from the current era. no_leader now defaults to yes for sides in
[scenario] and to no for sides in [multiplayer] so that the connect
engine now by default doesn't add a random mp leader to sides in sp
games.