Add a dynamic layout functionality to the game display
This adds dynamic layout functionality, which is invoked on a full
redraw and upon element creation. It also makes sure that there's no
attempt to do a surface restore on negative widhts or heights as
introduced support for in SDL2. Finally, there's a workaround for text
being clipped by 1-2 pixels in GUI1 buttons.