remove game_display instance form game_launcher
This fixes https://gna.org/bugs/?24139 so that there is always at most
in display object. display objects are the ingame/editor views which
use theme wml and show the map with its units. So the exist only during
the editor or ingame.

This removes some display_manager objects which had no effect because the
registered display was a dummy.