Add disabled flag to allow clearing of default hot-keys (Bugs #21983/#22218/#23981)
The original design of using a "null" command to indicate a disabled hot-key is ambiguous with regards to when to save a user hot-key to preferences as well as when a default hot-key should be flagged as disabled. So introduce a separate disabled flag to resolve the ambiguity.
Where the disabled flag is set to true, the hot-key should not be written to preferences unless it is a default hot-key.