move most dark forecast code to lua
This has multiple advantages:
1) It makes the code easier to understnad becasue we now can use local
variables and return values.
2) It makes the code easier to debug becasue we dont have to reload the
wml cache anymore for changes to have an effect
3) It makes loading the multiplayer wml faster, specialy dark forecasts
wml codes is now so short that we dont need to guard it into its own
preprocessor marco anymore.
4) It decreases the save file size.

This commit also does other improvements to the code, for example:
1) We don't create temporary units anymore just to quers the costs of
its unittype.
2) We now load the weather maps dynamicallyat runtimewinth
wesnoth.read_file, this removed the prorocessor paring time and savefile
size further.
3) When creating the new spaens wwe use temporary objects to decrease
the unit movement, instead of changing and resettign the unit stats
manually.

This also has a disadvantage:
1) It now requires all players to have this scenario installed, since it
it a mainline scenario i see no problems with it.