added border to dark forecast map
this adds borders to the 2p dark forecast survival map, it also removes
the big empty space in the upper third of the map.

This also renames the mask files to map files, becasue they are actually
just normal maps, and the [terrain_mask] just replaces the current map
with these.