Make the MAKE_ENUM Doxygen documentation actually sane
Otherwise, most of the contents get attached to the bad_enum_cast type
definition instead.

Note that I couldn't figure out why the documentation for the stream
input operator says that twml_exception is thrown when in debug-mode
when the code does not actually do that at this time (the relevant
function call is commented out).