Ensure that GUI1 widgets are drawn on a DRAW_ALL_EVENT
The DRAW_ALL_EVENT is only used in certain circumstances, when a full
redraw of all widgets/components is desired. The previous behaviour
relied on a DRAW-event following the DRAW_ALL_EVENT, which is not
something that can be relied upon. By calling draw() explicitly it
removes the reliance on the DRAW-event.