LoW S3: fixed incorrectly placed labels (bugs #22931, #23776)
I'm not sure why, but it seems the labels need to be initialized in prestart
if using [shift_labels] since I removed the explicit loading of that tag's lua
in a preload event. I'm also not sure why the shift values needed to be updated,
but everything seems to work now, which is what matters.