Make some transient messages transparent again
The transient messages transparency relies on the restorer
pattern. This enables it for certain transient messages, making sure
that they are drawn properly. Enabling it across all messages will
cause breakage.

This is a bit of a workaround, it should be changed to not rely on the
background restore functionality.