tlabel: align text to top of available space
In most cases, the widget size is the size of the text and looks no different
with this change. However, in cases where the label widget area is larger than
the text (such as in Unit Create), it resulted in text being rendered in the
middle of the available space, which looked weird.