Fix backstab always showing augmented damage in sidebar
Until now, [damage] (and possibly other) specials which relied on filters involving the opponent would display
in the sidebar as if the augmentation were always active (though when attacking, it worked correctly).
This is now fixed - such specials never show the augmentation in the sidebar now.

In addition, some minor code cleanup (initializing the weapon special adjacency filters outside the loop).