improve map::overlay()
we now read the config into objects at the start of that method, this means
that we don't have to do those lookups inside that loop anymore.

This also allows us to change that parameter of that function from config to
std::vector<overlay_rule> later when we want to use that
function from non-wml code.