Implement hotkey release handling.
Hotkeys are now triggered by a key/button release as well as a press.
This information is passed along as HOTKEY_EVENT_TYPE, which may be
HOTKEY_EVENT_PRESS, HOTKEY_EVENT_RELEASE, or HOTKEY_EVENT_REPEAT (for a
held key repeat).

Currently this should preserve the previous functionality. Existing
hotkeys respond to a press or repeat as they did before and ignore a
release.