disable c++14 build
Travis fails to build with c++14 because the default installed compiler (or rather the default libstdc++ from gcc 4.8) doesn't support it, i tried to fix it by changing the the install script to install gcc5 but somhow that failed too.

Since noone in our active dev team can fix this, we disable this for now. But i'd be useful if this was reenabled and fixed at some point.