Demote spurious network error message to info log level
The following message:

    20160516 16:59:30 error network: Destroing playturn_network_adapter with an non empty buffer, this means loss of network data

More often than not, this comes up when performing normal actions that
involve an unexpected but normal quit from a network game (quitting back
to lobby, closing the game), and the 'error' label makes it seem like a
serious issue when it isn't.

Also fixed typos in it.