Improve the control reassignment dialog when a player leaves (#654)
This moves the options for giving control to an ally/observer to the top of the list, labels them more descriptively, and prevents the dialog from appearing during linger mode if there is no next scenario. Also, dismissing the dialog by pressing esc no longer produces a network error, but instead simply defaults to setting the side to idle.