allow to write multiple configs in a network package.
It is now possible to write multiple config objects in a single package,
the server reads them then as one big config. Prevoulsy if people wanted
to do that they had to copy both configs into a new config which is
slow, specially if those configs are very big becasue they contain a
whole scenario or an era.