fix client not sending [leave_game]
this restores some codes that send the [leave_game] that was accidently lost when porting the client cost to boost asio.
I'm still unsure about whether it really makes sense to do it this way, specially that we send [leave_game] after (instead of before) calling enter_create_mode in line 501. (Which is was the pre-boost-asio code did aswell).