allow to special apceisl locations in the map_data
it is now possible to specify locations in the editor that can then be
read with wml (without the create scenario editor mode). These special
locations can be placed like starting positions, and can them be used in
the wml code like:
'location_id=<id>' in standard location filters or 'placement=<id>' in
[unit], i also intend to add a lua interface wesnoth.special_locations
later.

It's currently not possible to have multiple special  locations in the
same hex, the main reason is that i'm unsure about how to integrate that
in the editor ui.