select special location in editor palette after creating it
This removes the scroll_top() call in palette_manager::adjust_size()
becasue that made it impossible to scroll to the newly created item
after the "New Location Identifer" dialog was shown.

This changes most function in location_palette to use int instead of
size_t becasue "pos - num_visible_items() + 1" might result in a
negative number.

This also fixes a crash that happend when the "Go To" button was used on
an item that was no starting position.