AOI: Fix WML no unit for role, and other message edge conditions
In cases where there is no unit with the advisor role to deliver a message at the end of a scenario, a WML error appears on the screen.

While fixing these, other edge conditions were noted.

This changes end-of-scenario messages to select the oldest available using still alive to deliver the message.
It also supresses Erlornas' messages when having a conversation with no unit around to speak to.

Also, if none of the Mage apprentices of Linaera's are alive in the recall list, this supresses the WML error when one could not be recalled.

Finally, several scenarios simply ended when time ran out. Added Narration messages so all end with some bleak outlook for the future.