AOI Fix WML unit not found for role
My original goal was simply to suggest Linaera should be able to recruit Mage instead of using Erlornas' recruit list. But, in 1.13.4+dev, I hit some bugs which were not in 1.12. So, this is an omnibus which fixes those bugs.

One of those WML bugs was in S06 on victory where Linaera's student wants to speak but there is no surviving student. The fix for the advisor was to promote, recall or recruit. So that was what I did here. But, if Linaera can recruit here, why not let her recruit all the time? So fixing the bugs seems to lead to my original intention of allowing the player to use Linaera to recruit Mages, anyway.

1) In S05, using 1.13.4+dev, [teleport] for Linaera silently fails. Use [modify_unit] instead as a workaround. Added a TODO comment as a reminder to revert this. But it may be better if all four units (Linaera and her three student Mages) used the teleport arrival animation .. simultaneously, if possible .. I just don't know how to do this.

2) Traditionally, Linaera uses Erlornas' recruit list. Give each leader their own recruit list. Allow Linaera to recruit only Mage

3) Fix the WML unit not found for role errors when the advisor recalled (or recruited) at the begining of the scenario does not make it to the end. Promote a surviving unit as the new advisor. If no units survive, recall or recruit one.

4) In cases where the advisor was not used, messages at the end of a scenario would be silently dropped. Change those messages to the advisor and promote one, if needed.

5) In S01, time over event, Erlornas would always be the eldest surviving elf. The intent appears to be that another elf deliver the message of hopelessness. So change this to the advisor. This aligns the time over here to be more like that in other scenarios, where used.