Liberty Fix WML unit not found for role
In 1.13.4+dev I noted WML errors in S02 through S04 when attempting to locate a unit for use as an advisor.

I moved the macros over from AOI for locating, promoting and recalling an advisor. For this campaign, the unit types dynamically change; so I added a paramter which specifies the unit type to recruit if no suitable unit could be found. For this, I chose Thug and Thug_Peasant.

As with AOI, these macros select the eldest unit from the board or, of none exist there, from the recall list. This is a slight change from the original: my change allows the Footpad_Peasant to be an advisor, the original omitted this unit as a possibility. I did this so I could simply select a unit which was not one of the two heros and not worry about the specific unit types.

I also changed "Advisor" to "advisor" to align with AOI.

In S03 a [message] used "speaker=Advisor" and never appeared as a result. I changes this to "role=advisor" to have the message appear now that there is reliably an advisor available to deliver the message.

Finally, while testing, I ran into a but in 1.13.4+dev having to do with string-to-integer conversion. This affected S08, only. I changed the word 'FOREVER' to specify '99' turns, instead, to work around this bug. Since the scenario should end on turn 48 (or earlier), that is effectively the same. I can add a revert patch to this PR if a fix appears in the master; but it should not matter if this is never reverted.