LIBERTY Revised and improved fix for missed speaking parts and 'XML unit not found for role' errors on screen
The first attempt was too heavy-handed and did not work in all cases.

This version uses all ten recruitment times as possible advisors, and uses the same pattern everywhere. It also only appoints a new advisor when one is not available. It also uses a much lighter-weight macro to get the job done.

In S02, appoint an advisor from the units on the board. This advisor has a speaking line. If there are NO units on the board, appoint an advisor from the recall list; and the speaking line is silently skipped. If there are also no units in the recall list, do NOT emit the 'XML no unit for role' message on-screen. If the advisor's speaking line is missing, the flow of the conversation is still good.

In S03, simply recall the advisor from S02. If S02 didn't find one, we won't find one here and, if it did, it cannot have changed. This advisor has a speaking line.

In S04, the advisor could have died in S03. There won't be any units on the board, yet, so select one from the recall list. If none could be found, the speaking line is skipped. Again, the flow of the conversation is not adversely effected.

Of particular note is that the unit type names changed between S02 and S03 (in S02, after appointing the advisor). The sneeky bug-point here is the attributes of the units are copied, so the role had to be added to the copying code block.