DM S19 Defeat condition incorrect
Defeat should occur on the loss of the LAST undead available to Delfador. However, defeat occurs on the loss of ANY undead, UNLESS Delfador has EXACTLY TWO undead, counting the one in the process of dieing. The count was wrong, it should be two or more. 'Or more' is not supported, so 99999 (the default upper limit according to the Wiki) is used instead.