AToTB Fix errors
S01 "Rooting Out a Mage"

On Mordok's last breath, a speaking part goes to one of the units on the board. Some of the units are not considered, nor are unit advancements. In addition, there may be no units other than our leader, Arvith. These errors result in the WML error "no unit for role" to appear on-screen. Improve the storyline a bit by not selecting for the role but, instead, creating a loyal spearman to run in from the north to deliver this line, avoiding the entire mess.

On turn 10 a mercenary horseman delivers a line. Originally, this role was assigned at the start. If that unit died, the line would be omitted, which interrupts the flow of the script. Instead, assign the role when the line is neeeded and, if no horsement survive to deliver the role, give the player a free loyal horseman to deliver the line. Have the horseman come running in from the south-east corner.

The above to changes create new translatable strings for the unit names. Since these should rarely actually be translated, this should not create an issue. For both, the units run into the scene and move to be adjacent to our leader, Arvith.

S02 "The Chase"

On the last breath of the Dark Adept, there are speaking lines for a unit. If these are skipped the script makes no sense. At the beginning a speaker was assigned. If that speaker has died, another unit on the board is assigned the role. But, if there are no other units, the WML error "no unit for role" appears on the screen. Fix this by bringing in a loyal named spearman to deliver the lines. Try from the recall list and, if none are there, create a new one. This new speaker runs in from the right side of the board to join the leader.

In relation to this change, sumplified the WML a bit as well by removing an unneeded [NOT] tag.

This change creates a new translable string for the new loyal spearman. Again, this should not be an issue.