Clean up and fix logic
On [if] do not omit [then]

Fix the logic in S02 so, if we have a speaker, or assign a unit on the board as a speaker, it does not move.

In S02 if we need to recall or recruit a speaker, place it at the bottom edge and have it run up to the leader as if arriving late to the battle.