AToTB S02 Ensure a speaker exists during victory
The speaker is needed twice, and may have died in the meantime. Both times, we need him to give his lines so the conversation makes sense.

This changes the alternate speaker (during victory) from any unit on the board to a horseman. The line refers to 'riding' so this makes more sense, anyway.