AOI Fix missing speakers, use new [role] features
For each speaking part, make sure it is assigned to a unit. For parts where the conversation suffers if not delivered, recall a unit or recruit one if none are available. For parts where the missing the line does not appear to adversely effect the conversation, omit the lines if no unit can be found on the board at the time.