DM S14 S20 Fix bug: Kalenz takes over incorrectly
This bug causes several messages about unknown attributes to appear on stderr.

The first issue is the attributes are in the KALENZ macro, which appears in a [unit] tag.

Fixing this cascades to a new error on S20. The cause here is that Kalenz should take over as a unit on a no-leader scenario instead of being the leader. Copied solution from S13, where he first takes over.