DM S12 S19 Fix bug: Subsumed units cannot move
In S12, Delfador gains control of the Loyalists and, later, the Mages. In S19, he gains control of Lionel and his Loyalists.

This moves their turn up to round 1. Reset the units to full movement and attacks so they can move this turn; otherwise the units lose a full turn of movement.

This is most common with Arpus, the leader of the mages who, typically, cannot move even though her students can.

N.B. Only movement and attacks are affected. If the unit is injured, poisoned, slowed, etc., due to previous combat, it remains so. (Tested and verified in S12.)