Remove a performance optimization
The attack prediction code used to skip creating the "A slowed" or
"B slowed" plane if A or B was already slowed. As a result, after the
battle the "summary" array of the combatant had a different meaning from
usual: normally summary[0] means the HP distribution when the combatant is
not slowed, and summary[1] means the HP distribution when the combatant is
slowed. However, if the combatant was already slowed before the battle,
summary[0] was the HP distribution when the combatant is slowed and
summary[1] was empty.

This optimization didn't break anything, thanks to itself: because the
"A slowed" or "B slowed" plane wasn't used, the code fetched the
HP distribution from summary[0] instead of summary[1] and therefore
obtained the correct distribution.

However, my Monte Carlo simulation code requires that summary[0] and
summary[1] have their usual meanings. I decided that the optimization is
too hard to support, and removed it.