DM S07 Fix bug: Always early completion
The skeleton generators turned off on turn 12, always leading to an early-completion on turn 13. Don't turn them off. This allows normal turns-over completion after 14 turns, as designed.

Note than if the generators are taken out, and all skeletons killed, the player must wait for the next turn for victory. This can be rationalized by Delfador's statement indicating he is not sure his solution will work until the next turn, when no skeleton appears.

The use of an invalid tile to record the state of each generator caused errors generating the preview minimap when loading a turn.  Changing the tiles also caused visual artifacts. Eliminated the custom tiles and use event management to handle the generator state.