DM Fix bug: Subsumed sides cannot move
In S13 we subsume the Loyalists and Mages.

In S19 we subsume Lionel and his troops.

These occur in the midst of battle. The units may be damaged, poisoned, etc. We do not want to change that.

However, since their turn follows ours, they cannot move. Reset their moves to maximum so we can use them (or move them out of harm's way) immediately.