Fix crash by reintroduce splice.
This fixes a crash that would be caused by memory incorrectly being freed
when removed from a list. Relying on splice stops that from happening.

This also fixes an exit condition when deleting handlers as depth could
sometimes not be decremented.

Finally, checks are made to made sure that an sdl_handler is either
exclusively in the global context or a different context, ensuring
that they cannot be in both.