play_controller::check_victory(): use underscore in event name
Semantically this doesn't change anything because spaces in event names are
automatically replaced with underscores.

I just spent a considerable amount of time searching for code that fires
the enemies_defeated event, only to find that it's called
"enemies defeated", with a space, at the call site. Let's use the
underscored name everywhere to make the function easier to find.