Fix scenario end events not being run on debug-killing the enemy leader
If a developer used the "kill unit" debug command to end a scenario,
the scenario did something special on victory (e.g. showing messages)
and the scenario didn't explicitly set victory_when_enemies_defeated=no,
the special actions were skipped.

The cause was a reference with a too long lifetime.
Menu_handler::kill_unit() obtained a reference to the mouse position and
used it to send "last_breath" and "die" events. But if the mouse position
changed in between (for example, as a result of the last_breath event
doing something), the reference switched to point to the new mouse
position. As a result, the function sent the "die" event with incorrect
coordinates, filtering dropped the event, and the event handlers weren't
run.

Fixed by copying the mouse position instead.