Fix bug: libpng warning: iCCP: known incorrect sRGB profile
Some changes in libpng 1.6+ cause it issue warning or even not work correctly with the original HP/MS sRGB profile. You will notice this in stderr:

    libpng warning: iCCP: known incorrect sRGB profile

The old profile uses a D50 whitepoint, where D65 is standard. This profile is not uncommon, being used by Adobe Photoshop, although it was not embedded into images by default.

The simplest solution is simply to remove the embedded profile from your image. This can cause a slight shift in color *IF* have have a properly calibrated system, monitor, and software.