HttT S13 Add some random snow
The amount of snow is based upon how quickly you made it through S12 (Northern Winter). It is 1/3rd the coverage from S12, so it varries from 0 to 33% total coverage of the map, since we don't want to really bog down the player. At 1/3rd coverage, if S12 was won on the last possible turn the player here should only lose a turn or two getting to the exits.