HttT S05b Randomize temples
Formerly, to find Moremirmu, simply check (11,13) and, if not there, he is in (10,17). Oh HARD, there was a 1-in-2 chance of finding Xakae in (11,13), instead, and (9,15) was always empty. On NORMAL, Xakae is in (9,15).

Now, the odds of finding Moremirmu in a given temple are 1-in-3, as are the odds of finding Xakae unless playing on EASY where he is not present.