HttT S10 Clean/extend gryphons
The logic used some odd, hard-to-follow mathematics. But, in the end all it said was the Gryphons would put on a special display, but only if they had been attacked or killed by Robryn's troops, but not attacked or killed by Konrad's. Gave the mess the boot and put in clear logic.

Added some animations to the special display, and reduced some other functions.

The changes are a lot less scary than they may appear.