HttT S06 Logical Thieves
If Konrad has already captured a village in Elensefar it makes no sense for the thieves to offer help. Instead, just have them reinforce Konrad.

Seeing the control flow was hard so I rearranged the code, using macros to cut down on long sections between flow-points as well as eliminate repetition.

These changes look a lot more scary than they actually were; but there is a lot of motion so it can be hard to follow.