HttT S07 Fix bug: Units not hiding
Placing a duplicated unit causes it to drop into an adjacent hex and being missed by the object.

This only happens rarely, and almost always at the overlaped column. But it can happen anywhere. We'll just skip it and so what if we're a unit shy. Call it a very tiny, quite rare Easter Egg.