Damage calculation dialog: use existing battle prediction
The game already predicts the outcome of all possible battles when the
player opens the unit attack dialog, for the purpose of choosing the best
weapon to attack. Simply use that data in the damage calculation dialog
instead of recalculating it.

The change in mouse_handler::fill_weapon_choices() ensures that the damage
prediction will be present for all weapons. If bc is first copied to the
vector and bc::better_attack() is called only after that, the copy of bc in
the vector will never receive the damage prediction result.