Add new event handlers to a staging area
This reverts the semantical change in commit 61ccf2f that event handlers
registered as a result of an event received that event.

Compared with commit 3d8c29bc, this is a lower-level way to fix the ;
character appearing in the command console. This commit has potential to
fix more code that assumes that an event handler can't receive an event
that caused the handler to be registered in the first place. On the other
hand, it is no longer possible for event handlers to receive such events
even if developers want that.