AToTB S02 Agressive kidnappers
This combines two changes.

First, it uses the sighted event instead of a moveto tripwire to find the kidnappers.

Second, instead of passively waiting for a target to come into range, the kidnappers behave as more normal AI, but ignoring keeps and devalueing villages.

This is done by moving the kidnappers to their own side, which remains hidden until they are discovered.