TSG S05 Let Urza Afala roam free!
There is no reason to stop his moving about. Separating his sighted event and finding the bandit encampment eliminated any problems with his movement. This was a todo comment for when this was fixed. Actually, I think the real problem was the inapproriate sighting of the bandits when we found the lich.