TSG S01 Fix bug: Urza Mathin may die first
In the unlikely event the player manages to kill Urza Mathin before visiting the river fort, Sir Gerrick will not appear. This is bad for the victory conversation and worse for the rest of the campaign.

I thought long and hard about how to fix this and decided to take my cue from the initial objectives which show two items. So, I made it so you have to do BOTH to win the scenario. If you kill Urza Mathin first, you still need to go find Sir Gerrick. I only needed to adjust one line to make it make sense, and it works better than simply having Sir Gerrick pop up out of nowhere.