TSG S06a Fix bug: Units in poor shape
When we next see Sir Gerrick and the units he has on the map will be in S09a. While Sir Gerrick will be fine, the units will appear to be in poor shape in the recall list. Let's heal everyone, first, so we're sure everyone is in fighting shape when they return.