Menu Button: increase min width
This serves the immediate purpose of making the Resolutions Menu Button in Prefs equal the size of
its dropdown, but it also ensures there's more space for lists with small initial entries but longer
choices. Obviously, horizontal_grow = true could be used in any dialogs where this is a case, but this
provides a safer minimum should that not be desired.