Load Game: display all human-controlled leaders
This also includes a fix for leader images sometimes not being shown in the dialog.
If the save_index was deleted or altered at a time other than when creating a save,
the image path would be empty in the game load dialog. This ensures the image path
is only saved in binary-path independent form if the path isn't empty. In that case,
the plain image path from the savefile is saved. Either way, the image path is checked
in-dialog, first plain, then as binary-path independent. If both fail, a fallback image
is shown.