Load Game: fixed game list being empty with long filenames on small resolutions
For some reason, without horizontal_scrollbar_mode = "never", the game didn't utilize the
window scrollbar and instead simply failed to place the listbox content grid. This resolves that,
as well as reducing the the filename truncate size so names fit fine on 800x600 without said
window scrollbar. Also, it ensures a horizontal scrollbar won't appear in the leaders list.