Fix bug: user_team_name is wrong
When pressing End Scenario and proceeding to the next scenario, the user_team_name shown on Menu|Status Table is incorrect.

If you Load (from the title screen) the saved scenario-start, the correct user_team_name appears.

This is caused by the random-shuffle feature for multi-player scenario. This fixes it for Single-Player campaigns and tries not to break it too badly for randomized Multi-Player campaigns.