Fix crash in GUI2 multiplayer test
Because the join script attempts to join the game immediately after it has
been found, it tried to join the game before it had been added to the
games_ vector (from games_by_id_).

In addition to that, the check for if there is such a game had an
off-by-one error.

The two bugs, in combination, caused the game to index an empty vector.
Even if the STL implementation doesn't immediately crash the game, it's
going to crash later when it tries to interpret garbage data as the game
information.

Fixed by making the game update the games_ vector immediately after
fetching games from the server, and also fixing the off-by-one error.