HttT S19a Fix bug: Paladin can use sword
An earlier change from blade to arcane for the Paladin's Holy Sword made the Flaming Sword ineffective and he was removed from the list of unit types allowed to take the sword.

This re-instates Paladin by changing the effect filter. I checked all CURRENT units and minimized the filter so the Flaming Sword only effects the swords and the sabre Li'Sar uses.