Change most vector<bool> to dynamic_bitset
A few cases were skipped, most of which involved 2D bitsets (eg shroud)
or uses that could not be easily translated (eg GUI1 multimenu).

In a few cases, code needed to be updated for the different API that the
dynamic_bitset provides, but in most cases, this change is entirely
transparent.

There are probably a few unnecessary dynamic_bitset.hpp includes.

One use of vector<bool> was replaced instead with std::bitset, since
the dynamic size was not required.