Check for occupied tunnel exits at teleport map creation
Previously this was done in the move execution, which resulted in:
1. own and allied units on tunnel/teleport exits effectively acting as
ambushers and
2. unit reaches and shortest paths not being consistent with actually
possible moves.
Both issues occurred even without fog being active.

The former was unintuitive at best, while the latter could cause
serious problems, for example with AI code. By moving the check for
units on tunnel exits to the teleport map creation, both these issues
are taken care of simultaneously.

Only visible units are considered. In the presence of fog, enemy units
and units from allies without shared vision continue to interrupt the
move as expected.